🏗️ Split player state store types out into shared
